replace default int-types with safe fix length types in ZCodeHeader

+++ code/TweeZcodeCompiler/main/ZCodeHeader.cpp
@@ -16,7 +16,7 @@ using namespace std;
 #define STANDARD_REVISION_MAIN 1 // revision number of supperted document, here
 #define STANDARD_REVISION_SUB 1  // 1.1 (STANDARD_REVISION_MAIN.STANDARD_REVISION_SUB)
 
-vector<bitset<8>>*ZCodeHeader::getHeaderBits() {
+vector<bitset<8>> *ZCodeHeader::getHeaderBits() {
     if (!fileLengthSet) {
         cout << "No file size specified!" << endl;
         throw;
@@ -44,7 +44,7 @@ vector<bitset<8>>*ZCodeHeader::getHeaderBits() {
     setShortVal(screenWidthUnits, headerBits);       // Hex 22 - 23
     setShortVal(screenHeightUnits, headerBits);      // Hex 24 - 25
     headerBits->push_back(fontWidthUnits);           // Hex 26
-    headerBits-> push_back(fontHeightUnits);         // Hex 27
+    headerBits->push_back(fontHeightUnits);         // Hex 27
     setShortVal(routinesOffset, headerBits);         // Hex 28 - 29
     setShortVal(staticStringsOffset, headerBits);    // Hex 2A - 2B
     headerBits->push_back(defBackgroundColor);       // Hex 2C
@@ -66,14 +66,14 @@ vector<bitset<8>>*ZCodeHeader::getHeaderBits() {
 /* helper method to calculate length and to
  * avoid not setting these values before calling getHeaderBits()
  * */
-void ZCodeHeader::setFileLength(unsigned int length, unsigned short checksum) {
-    int len = length / 8;
+void ZCodeHeader::setFileLength(uint64_t length, uint16_t checksum) {
+    uint64_t len = length / 8;
 
-    if (len > numeric_limits<unsigned short>::max()) {
+    if (len > numeric_limits<uint16_t>::max()) {
         cout << "File too large" << endl;
         throw;
     } else {
-        fileLength = len;
+        fileLength = (uint16_t) len;
         fileChecksum = checksum;
         fileLengthSet = true;
     }
@@ -82,14 +82,14 @@ void ZCodeHeader::setFileLength(unsigned int length, unsigned short checksum) {
 /* helper method to calculate routines offset and to
  * avoid not setting this value before calling getHeaderBits()
  * */
-void ZCodeHeader::setRoutinesOffset(unsigned int offset) {
-    int off = offset / 8;
+void ZCodeHeader::setRoutinesOffset(uint64_t offset) {
+    uint64_t off = offset / 8;
 
-    if (off > numeric_limits<unsigned short>::max()) {
+    if (off > numeric_limits<uint16_t>::max()) {
         cout << "Routine offset too large" << endl;
         throw;
     } else {
-        routinesOffset = off;
+        routinesOffset = (uint16_t) off;
         routinesOffsetSet = true;
     }
 }
@@ -97,14 +97,14 @@ void ZCodeHeader::setRoutinesOffset(unsigned int offset) {
 /* helper method to calculate static string offset and to
  * avoid not setting this value before calling getHeaderBits()
  * */
-void ZCodeHeader::setStaticStringsOffset(unsigned int offset) {
-    int off = offset / 8;
+void ZCodeHeader::setStaticStringsOffset(uint64_t offset) {
+    uint64_t off = offset / 8;
 
-    if (off > numeric_limits<unsigned short>::max()) {
+    if (off > numeric_limits<uint16_t>::max()) {
         cout << "Routine offset too large" << endl;
         throw;
     } else {
-        staticStringsOffset = off;
+        staticStringsOffset = (uint16_t) off;
         staticStringsOffsetSet = true;
     }
 }
@@ -160,8 +160,8 @@ void ZCodeHeader::setFlags2(std::vector<std::bitset<8>> *header) {
 }
 
 // splits short value up to 2 bytes and pushes them into the headerBits
-void ZCodeHeader::setShortVal(unsigned short val, vector<bitset<8>> *header) {
-    bitset<16> shortVal (val);
+void ZCodeHeader::setShortVal(uint16_t val, vector<bitset<8>> *header) {
+    bitset<16> shortVal(val);
     bitset<8> firstHalf, secondHalf;
 
     for (size_t i = 0; i < 8; i++) {
@@ -169,7 +169,7 @@ void ZCodeHeader::setShortVal(unsigned short val, vector<bitset<8>> *header) {
     }
 
     for (size_t i = 8; i < 16; i++) {
-        firstHalf.set(i-8, shortVal[i]);
+        firstHalf.set(i - 8, shortVal[i]);
     }
 
     header->push_back(firstHalf);
